# go-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST que calcula la factorización QR de una matriz y obtiene estadísticas desde Node.js.",
        "title": "QR Matrix API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/matrix/qr": {
            "post": {
                "description": "Recibe una matriz numérica, calcula Q y R con Gonum y obtiene estadísticas de la API Node.js",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "matrix"
                ],
                "summary": "Calcula la factorización QR de una matriz",
                "parameters": [
                    {
                        "description": "Matriz de entrada",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MatrixRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.MatrixResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "middleware.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.MatrixRequest": {
            "type": "object",
            "properties": {
                "matrix": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "number",
                            "format": "float64"
                        }
                    }
                }
            }
        },
        "models.MatrixResponse": {
            "type": "object",
            "properties": {
                "q": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "number",
                            "format": "float64"
                        }
                    }
                },
                "r": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "number",
                            "format": "float64"
                        }
                    }
                },
                "statistics": {
                    "$ref": "#/definitions/models.Statistics"
                }
            }
        },
        "models.Statistics": {
            "type": "object",
            "properties": {
                "average": {
                    "type": "number"
                },
                "isDiagonalQ": {
                    "type": "boolean"
                },
                "isDiagonalR": {
                    "type": "boolean"
                },
                "max": {
                    "type": "number"
                },
                "min": {
                    "type": "number"
                },
                "sum": {
                    "type": "number"
                }
            }
        }
    }
}